refactor: extract shared atomicWriteText utility (#415)
Extract the duplicated atomic-write-to-temp-then-rename method from
three YAML repositories into a shared top-level function in FileUtil.kt.

- YamlPlayerRepository: removed 20-line private method
- YamlGuildRepository: removed 17-line private method
- YamlWorldStateRepository: removed 17-line private method

All three now call the shared atomicWriteText(path, contents) function.
Net reduction of 38 lines (34 new + 72 removed).
