+++ src/main/kotlin/dev/ambon/persistence/YamlGuildRepository.kt
@@ -8,12 +8,7 @@ import dev.ambon.domain.guild.GuildRank
 import dev.ambon.domain.guild.GuildRecord
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
-import java.io.IOException
-import java.nio.file.AtomicMoveNotSupportedException
-import java.nio.file.Files
 import java.nio.file.Path
-import java.nio.file.StandardCopyOption
-import java.nio.file.StandardOpenOption
 import kotlin.io.path.createDirectories
 import kotlin.io.path.deleteIfExists
 import kotlin.io.path.exists
@@ -92,24 +87,6 @@ class YamlGuildRepository(
     private fun write(record: GuildRecord) {
         val path = pathFor(record.id)
         atomicWriteText(path, mapper.writeValueAsString(GuildDto.from(record)))
-    }
-
-    private fun atomicWriteText(
-        path: Path,
-        contents: String,
-    ) {
-        try {
-            path.parent?.createDirectories()
-            val tmp = Files.createTempFile(path.parent, path.fileName.toString(), ".tmp")
-            Files.writeString(tmp, contents, StandardOpenOption.TRUNCATE_EXISTING)
-            try {
-                Files.move(tmp, path, StandardCopyOption.REPLACE_EXISTING, StandardCopyOption.ATOMIC_MOVE)
-            } catch (_: AtomicMoveNotSupportedException) {
-                Files.move(tmp, path, StandardCopyOption.REPLACE_EXISTING)
-            }
-        } catch (e: IOException) {
-            throw PlayerPersistenceException("Failed to write guild file atomically: $path", e)
-        }
     }
 }
