+++ src/main/kotlin/dev/ambon/persistence/YamlWorldStateRepository.kt
@@ -7,12 +7,7 @@ import com.fasterxml.jackson.module.kotlin.readValue
 import io.github.oshai.kotlinlogging.KotlinLogging
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
-import java.io.IOException
-import java.nio.file.AtomicMoveNotSupportedException
-import java.nio.file.Files
 import java.nio.file.Path
-import java.nio.file.StandardCopyOption
-import java.nio.file.StandardOpenOption
 import kotlin.io.path.createDirectories
 import kotlin.io.path.exists
 import kotlin.io.path.readText
@@ -52,22 +47,4 @@ class YamlWorldStateRepository(
                 log.error(e) { "Failed to save world state to $stateFile" }
             }
         }
-
-    private fun atomicWriteText(
-        path: Path,
-        contents: String,
-    ) {
-        try {
-            path.parent?.createDirectories()
-            val tmp = Files.createTempFile(path.parent, path.fileName.toString(), ".tmp")
-            Files.writeString(tmp, contents, StandardOpenOption.TRUNCATE_EXISTING)
-            try {
-                Files.move(tmp, path, StandardCopyOption.REPLACE_EXISTING, StandardCopyOption.ATOMIC_MOVE)
-            } catch (_: AtomicMoveNotSupportedException) {
-                Files.move(tmp, path, StandardCopyOption.REPLACE_EXISTING)
-            }
-        } catch (e: IOException) {
-            throw RuntimeException("Failed to write file atomically: $path", e)
-        }
-    }
 }
